Gyroscope RV Sensor
Created a Rotation Vector Sensor which uses only accel and gyro

Note - In lines 288-290 of src/sensor/rotation_vector/fusion_utils/
orientation_filter.cpp, the fusion backend does not update the quaternion
if acceleration and gyroscope values are higher than a very low threshold.

While this may improve performance in high noise situations, doing so
causes the sensor to not update its values in most common-use scenarios.

I would hence recommend to remove the 3 if conditions in the lines above.

Change-Id: I0b96e6235c52f76195529b918bd54e29bf946631
Signed-off-by: akhilkedia94 <akhil.kedia@samsung.com>