Add 'Provides' to sensord-dummy package
Since the library name was changed,
(libsensor-genuine -> libsensor, libsensor -> libsensor-dummy)
the 'provides' of each package were also changed:
    sensord package is now providing 'libsensor.so.<major version>'.
    sensord-dummy package does not provide 'libsensor.so.<major version>' any more.
In a profile that uses only the sensord-dummy package, other packages
can require 'libsensor.so.<major version>'.
In this situation, rpm installer brings sensord package, which has
'libsensor.so.<major version>', even the profile does not use sensord package.
To fix this situation, 'libsensor.so.<major version>' was added to the
provides list of sensord-dummy.

Change-Id: Iec2c08088da1ca50b0d5421c6169f72da78e9ab7
Signed-off-by: SangYoun Kwak <sy.kwak@samsung.com>