Modify the requirement of devel package
The "devel" package gives hint about libraries to use the "sensor" package
to the other packages.
Before this commit, the "devel" package required the "sensor-dummy"
package. It was valid when the "sensor-dummy" has "libsensor.so.4".
But now, the "sensor-dummy" has no real "libsensor.so.4" but has
"libsensor-dummy.so" so the other packages which requires the package
"sensor" will require "libsensor-dummy.so", which is not correct.
Modifying the requirement of devel package as the "sensord" will fix
this.

Change-Id: I25eaa772bcfd7edbee46aaf823adfef36630f70e
Signed-off-by: SangYoun Kwak <sy.kwak@samsung.com>