packaging: Modify the installation of sensord and sensord-dummy
Previously, sensor-dummy was installed with sensord due to the
providence of the name 'libsensor.so.<major version>', which is a
reference for expanding the "package dependency".
In this patch, sensord will be installed independently and it will
provide the name "libsensor.so.<major version>". According to this,
sensor-dummy will be installed with sensord due to the providence of the
name.

Since the sensord daemon is not necessary on the target which uses
sensord-dummy, the post script of sensord-dummy will remove the files
which are installed by the sensord package.

The sensord-dummy package is not required explicitly in the tv profile.
(previously, sensord-dummy was required by the name "libsensor.so.<major
 version>", it was not necessary to require sensord-dummy explicitly)
Thus, this patch should be applied with the modification of
.ks(kickstart) file.
("sensord-profile_tv" which provided by the sensord-dummy can be used)

Change-Id: I0c56af9f6e9f35a35fc0b3b5074400d94ebd5ebe
Signed-off-by: SangYoun Kwak <sy.kwak@samsung.com>